tests with error 200 is strig
